Check "lights" exists in group
When creating an empty group in Hue app the Hue app will crash and the created group will look like the following in config.json:

'3': {'action': {'on': False},
       'class': 'Living room',
       'name': 'Test',
       'state': {'all_on': False, 'any_on': False},
       'type': 'Room'}}

Next a syncWithLights is called it will fall into the exception because the "Lights" property doesn't exist for this group.

Sadly it doesn't fix the empty group problem in the Hue app.