Update remoteApi.py
Update remoteApi to get it working even if you listen on a single interface.
Just renamed 'localhost' to HOST_IP, because with 'localhost' all requests are made to loopback interface, which of course is not the one we need.
In order to do this you need to pass HOST_IP as argument. It's not so neat, but it works. A solution could be moving remoteApi function to HueEmulator3.py or a complete restyling into a class-based application.

PS: HueEmulator3.py is being updated too to match this new argument.