Initial WLED Support
Using mdns(zeroconf library) to discover a WLED device and JSON to send commands(maybe UDP streaming sometime) to it we can use it in the Hue app as a rgb bulb.
It's not heavily tested at the momemt and there is a lot of room for improvements if the community is interested in such a feature.