Add support for setting color temperature in scenes (#783)
This fixes a bug that has existed for a while in the V2 API where Hue scenes that configure the color temperature have not been able to be created. This tends to apply a lot to the default Hue scenes that are provided in the app, since many of them are designed to operate without color to make it more compatible across devices. The result has been that only the brightness and on state has been saved for these scenes, resulting in unexpected and inconsistent results when trying to apply the scene.

This is fixed by persisting the color temperature value for each light when it is present in the scene.