Rationalize logging (#846)
Currently all logging information is written to stderr (the python logging default),
but this means systemd-journald interprets every log message as an error and sets
the log level for messages in the journal accordingly. This patch leaves only
warnings and errors logged to stderr, all other messages are logged to stdout.