Update Group.py
Fix for KeyError: 'bri'
The error occurred because some lights in a group did not have the bri (brightness) key in their state. To resolve this, the .get() method is used to provide a default value of 0 when bri is not available:

bri = bri + light().state.get("bri", 0)
This ensures that the program doesn't fail when the bri key is missing.